Products Done

+++ products/src/controllers/products.controller.js
@@ -66,12 +66,13 @@ export const createProduct = (req, res) => {
   })
 }
 
+// ...
+
 export const updateProduct = async (req, res) => {
   const productId = req.params.id
   const product = await Product.findById(productId)
   const publicUrl = product.publicUrl
   try {
-    const product = await Product.findById(productId)
     if (!product) {
       return res.status(404).json({ message: 'Producto no encontrado' })
     }
@@ -93,8 +94,10 @@ export const updateProduct = async (req, res) => {
         if (req.body.category) product.category = req.body.category
         if (req.body.stock) product.stock = req.body.stock
         await product.save()
-        await cloudinary.uploader.destroy(publicUrl)
-        fs.unlinkSync(req.file.path)
+        if (req.file) {
+          await cloudinary.uploader.destroy(publicUrl)
+          fs.unlinkSync(req.file.path)
+        }
         res.json(product)
       } catch (error) {
         res.status(500).json({ message: 'Error al actualizar el producto' })
@@ -104,6 +107,47 @@ export const updateProduct = async (req, res) => {
     res.status(500).json({ message: 'Error al buscar el producto' })
   }
 }
+
+// ...
+
+// export const updateProduct = async (req, res) => {
+//   const productId = req.params.id
+//   const product = await Product.findById(productId)
+//   const publicUrl = product.publicUrl
+//   try {
+//     const product = await Product.findById(productId)
+//     if (!product) {
+//       return res.status(404).json({ message: 'Producto no encontrado' })
+//     }
+//     upload(req, res, async (err) => {
+//       try {
+//         if (err) {
+//           return res.status(400).json({ message: 'Error al cargar la imagen' })
+//         }
+//         if (req.file) {
+//           const result = await cloudinary.uploader.upload(req.file.path, {
+//             folder: 'Prueba'
+//           })
+//           product.image = result.secure_url
+//           product.publicUrl = result.public_id
+//         }
+//         if (req.body.name) product.name = req.body.name
+//         if (req.body.description) product.description = req.body.description
+//         if (req.body.price) product.price = req.body.price
+//         if (req.body.category) product.category = req.body.category
+//         if (req.body.stock) product.stock = req.body.stock
+//         await product.save()
+//         await cloudinary.uploader.destroy(publicUrl)
+//         fs.unlinkSync(req.file.path)
+//         res.json(product)
+//       } catch (error) {
+//         res.status(500).json({ message: 'Error al actualizar el producto' })
+//       }
+//     })
+//   } catch (error) {
+//     res.status(500).json({ message: 'Error al buscar el producto' })
+//   }
+// }
 
 export const deleteProduct = async (req, res) => {
   try {
